moved to separate status files. removed json feature.
