+++ internal/pkg/warewulfd/runtime.go
@@ -1,7 +1,6 @@
 package warewulfd
 
 import (
-	"html/template"
 	"net/http"
 	"strconv"
 	"strings"
@@ -13,115 +12,6 @@ import (
 	"github.com/hpcng/warewulf/internal/pkg/util"
 	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
 )
-
-// TODO: move to separate file?
-const nodeStatusHtmlTemplate = `
-<!DOCTYPE html>
-<html>
-<head>
-<style>
-table {
-  font-family: arial, sans-serif;
-  border-collapse: collapse;
-  width: 100%;
-}
-
-td, th {
-  border: 1px solid #dddddd;
-  text-align: left;
-  padding: 8px;
-}
-
-tr:nth-child(even) {
-  background-color: #dddddd;
-}
-</style>
-</head>
-<body>
-    <h1>{{.PageTitle}}</h1>
-    <table>
-      <tr>
-        <th>Node</th>
-        <th>Cluster</th>
-        <th>last seen (s)</th>
-      </tr>
-        {{range .HtmlBody}}
-            <tr>
-                <td>{{.Node}}</td>
-                <td>{{.Cluster}}</td>
-                <td>{{.LastSeen}}</td>
-            </tr>
-        {{end}}
-    </table>
-</body>
-</html>`
-
-var NodeInfoDB = make(NodeList)
-
-type TemplData struct {
-	Node     string
-	Cluster  string
-	LastSeen int64
-}
-
-type PageData struct {
-	PageTitle string
-	HtmlBody  []TemplData
-}
-
-func NodeStatusSend(w http.ResponseWriter, req *http.Request) {
-	w.Header().Set("Server", "Warewulf status information")
-	// query handling
-	q := req.URL.Query()
-	if q.Has("test") {
-		count, err := strconv.Atoi(q["test"][0])
-		if err == nil {
-			NodeInfoDB.AddTestNodes(count)
-		}
-	}
-	if q.Has("json") {
-		NodeInfoDB.JsonSend(w)
-		return
-	}
-	// human readable
-	maxLines := len(NodeInfoDB)
-	if q.Has("limit") {
-		l, err := strconv.Atoi(q["limit"][0])
-		if err == nil {
-			maxLines = l
-		}
-	}
-
-	// Make and parse the HTML template
-	t, err := template.New("webpage").Parse(nodeStatusHtmlTemplate)
-	if err != nil {
-		http.Error(w, err.Error(), 500)
-		return
-	}
-
-	data := PageData{
-		PageTitle: "node status info",
-		HtmlBody:  []TemplData{},
-	}
-
-	// body
-	now := time.Now().UTC().Unix()
-	sorted := NodeInfoDB.Sort()
-	i := 0
-	for _, n := range sorted {
-		i++
-		if i > maxLines {
-			break
-		}
-		data.HtmlBody = append(data.HtmlBody, TemplData{n.Id.Get(), n.ClusterName.Get(), now - n.LastSeen})
-	}
-	// render and write
-	err = t.Execute(w, data)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
 
 func RuntimeOverlaySend(w http.ResponseWriter, req *http.Request) {
 	conf, err := warewulfconf.New()
@@ -172,8 +62,7 @@ func RuntimeOverlaySend(w http.ResponseWriter, req *http.Request) {
 		return
 	} else {
 		daemonLogf("REQ:   %15s: %s\n", n.Id.Get(), req.URL.Path)
-		n.LastSeen = time.Now().UTC().Unix()
-		NodeInfoDB[n.Id.Get()] = &n
+		NodeInfoDB[n.Id.Get()] = &NodeStatus{&n, time.Now().UTC().Unix()}
 	}
 
 	if n.RuntimeOverlay.Defined() {
